chore(worker): consolidate copy-pasted smoke-harness helpers into smoke_support (sc-8866)
Extract the byte-identical per-smoke-file helper boilerplate — env_or (the
empty-filtering env read unified in sc-8924), the RGB8 degenerate-decode floor
checks (image_mean/image_std/is_all_zero), and save_png — into one shared
#[cfg(test)] mod smoke_support, imported by every real-weight smoke harness
(the *_mlx_smoke.rs macOS lane, the *_gpu_smoke.rs off-Mac candle lane, and
footprint_measure.rs).

smoke_support is gated on the SUPERSET of both smoke lanes
(test AND (macos OR the off-Mac candle build)) so it compiles exactly where a
smoke that imports it does. Each smoke keeps only its own engine id, repo/tier/
sentinel resolution, and request recipe — nothing about what any smoke asserts
or which tests run changes; this is a pure test-support reorg. The two prior
image_std variants (self-contained vs calling image_mean) are numerically
identical, so consolidating onto the shared image_mean-backed version is
behavior-preserving.

is_all_zero is #[allow(dead_code)]: only the macOS packed-tier smokes assert
the all-zero floor, so it is legitimately unused on the candle-only lane —
allow rather than a cfg gate keeps it available to any future smoke either lane.
